Backend/requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
